configuration fixes
