+++ app/server.py
@@ -1,3 +1,4 @@
+import os
 from app import app, application
 from .layouts.main_layout import get_main_layout
 
@@ -13,4 +14,4 @@ if __name__ == "__main__":
     # Setting debug to True enables debug output. This line should be
     # removed before deploying a production app.
     application.debug = True
-    application.run(host="0.0.0.0")
+    application.run(host="0.0.0.0",port=8050)
