+++ requirements.txt
@@ -36,5 +36,4 @@ tzdata==2023.3
 urllib3==1.26.16
 Werkzeug==2.3.4
 flask-dynamodb-sessions==0.1.8
-simplejson
-awscli
+simplejson==3.19.1
